pyproject.toml
[tool.poetry]
name = "git-hours"
version = "0.1.2"
description = "Estimate time spent on a git repository "
authors = ["Stefane Fermigier <sf@abilian.com>"]
license = "GPL"
readme = "README.md"
packages = [{include = "git_hours", from = "src"}]


[tool.poetry.dependencies]
python = "^3.10"
pydash = "^5.1.1"
devtools = "^0.9.0"
GitPython = "^3.1.28"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"

[tool.poetry.scripts]
git-hours = 'git_hours.main:main'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
